pyproject.toml
[tool.poetry]
name = "eduport"
version = "0.1.0"
description = ""
authors = ["vishnu2044 <vishnunarayanan2044@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
django-cors-headers = "^4.4.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
